refact: the getTotalPrice convert to BigDecimal.

+++ src/main/java/com/codecool/shop/dao/CartDao.java
@@ -1,11 +1,8 @@
 package com.codecool.shop.dao;
 
 import com.codecool.shop.model.CartProduct;
-import com.codecool.shop.model.Supplier;
-import com.codecool.shop.model.Product;
-import com.codecool.shop.model.ProductCategory;
 
-import java.util.HashMap;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CartDao {
@@ -18,6 +15,6 @@ public interface CartDao {
 
     List<CartProduct> getAll();
 
-    Integer getTotalPrice();
+    BigDecimal getTotalPrice();
 
 }
